// src/interfaces/Errors.interface.ts
export default interface IErrors {
  name?: string;
  email?: string;
  password?: string;
  password2?: string;
  badRequest?: string;
  notWorking?: string;
  noProfile?: string;
  noUser?: string;
  noPost?: string;
  noPosts?: string;
  noProfiles?: string;
  notAuthorized?: string;
  notLiked?: string;
  commentNotExits?: string;
  alreadyLiked?: string;
  userId?: string;
}